requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
torch==1.12.1
pillow==9.3.0
ultralytics==8.0.56
clip==1.0
nest-asyncio==1.5.6